bugfixes, update gist when post changes

+++ app/models/post.rb
@@ -19,13 +19,17 @@ class Post < Mist::GitModel
   attribute :published_at
   attribute :gist_id
   
-  before_save :update_gist_if_necessary
+  after_validation :update_gist_if_necessary
   after_initialize :load_code_examples_from_gist
-  
+  after_destroy :destroy_gist
   
   def title=(value)
     self.id = permalink(value)
     attributes[:title] = value
+  end
+  
+  def content=(c)
+    attributes[:content] = c.gsub(/\r/, "")
   end
   
   def gist
@@ -37,7 +41,12 @@ class Post < Mist::GitModel
           nil
         end
       else
-        ActiveGist.find(gist_id)
+        begin
+          ActiveGist.find(gist_id)
+        rescue RestClient::ResourceNotFound
+          self.gist_id = nil
+          nil
+        end
       end
     end
   end
@@ -46,16 +55,26 @@ class Post < Mist::GitModel
     GitHub::Markup.render("#{title}.markdown", content_with_embedded_gists).html_safe
   end
   
+  def content_as_html_preview
+    # just take to the first blank line -- that's probably the first paragraph
+    # TODO make this smarter by including more than 1 paragraph if it's short, or by omitting headers
+    first_paragraph = /\A(.+?)(\n\n|\n    |\z)/m.match(content.gsub(/\r/, ''))
+    GitHub::Markup.render("#{title}.markdown", first_paragraph[1]).html_safe
+  end
+  
   def content_with_embedded_gists
-    return content.dup unless gist && gist.persisted?
+    # by using gist_id directly we can avoid hitting the Gist API every time the
+    # post is rendered.
+    
+    return content.dup if gist_id.blank?
     
     template = '<script src="https://gist.github.com/__ID__.js?file=__FILENAME__"></script>'
-    template['__ID__'] = gist.id.to_s
+    template['__ID__'] = gist_id.to_s
     
     content.dup.tap do |result|
       # process last example first, so that changes to result don't taint offsets
       code_examples.reverse.each do |example|
-        result[example.offset] = template.sub(/__FILENAME__/, example.filename)
+        result[example.offset] = template.sub(/__FILENAME__/, example.filename) + "\n"
       end
     end
   end
@@ -69,7 +88,6 @@ class Post < Mist::GitModel
   end
   
   def load_code_examples_from_gist
-    # p gist
     if gist && gist.persisted?
       self.content = self.content.dup.tap do |result|
         code_examples.reverse.each do |example|
@@ -83,24 +101,37 @@ class Post < Mist::GitModel
     end
   end
   
+  # Assigns the file contents of each file in the gist according to what's found in
+  # #content. Does not save the gist. Returns the list of files themselves.
+  def set_gist_file_contents
+    gist.files.keys.each { |filename| gist.files[filename] = nil }
+    code_examples.each do |example|
+      gist.files[example.filename] = { :content => example }
+    end
+    gist.files
+  end
+  
   def update_gist_if_necessary
+    return unless errors.empty?
+    
     if has_code_examples?
-      # mark all files for deletion, we'll effectively undo this below
-      gist.files.send(:hash).each { |key, info| gist.files[key] = nil }
+      set_gist_file_contents
       
-      code_examples.each_with_index do |example, index|
-        example_file = gist.files[example.filename] ||= {}
-        example_file[:content] = example
-      end
+      # mark any additional files for deletion
+      # gist.files.send(:hash).each { |key, info| gist.files.send(:hash).delete(key) unless filenames.include?(key) }
       
       if gist.changed?
         errors.add(:gist, "could not be saved: #{gist.errors.full_messages.join('; ')}") unless gist.save
       end
-      
+
       self.gist_id = gist.id
     else
       # no code examples, delete gist
       gist.destroy if gist && gist.persisted?
     end
+  end
+  
+  def destroy_gist
+    gist.destroy if gist && gist.persisted?
   end
 end
